+++ packages/cms/lib/modules/openstad-global/lib/fields.js
@@ -145,11 +145,48 @@ module.exports = [
     label: 'Hide the site title in the header?',
     def: true
   },
+
+  {
+    type: 'select',
+    name: 'analyticsType',
+    label: 'Analytics type',
+    def: 'google-analytics-old-style',
+    choices: [
+      {
+        value: 'none',
+        label: "No analytics",
+      },
+      {
+        value: 'google-analytics-old-style',
+        label: "Google Analytics old style (with a property like UA-xxxxx)",
+        showFields: ['analyticsIdentifier']
+      },
+      {
+        value: 'google-analytics-new-style',
+        label: "Google Analytics new style (with a property like G-xxxxx)",
+        showFields: ['analyticsIdentifier']
+      },
+      {
+        value: 'custom',
+        label: "Custom: use a custom codeblock",
+        showFields: ['analyticsCodeBlock']
+      }
+    ]
+  },
+
   {
     type: 'string',
-    name: 'analytics',
-    label: 'Google Analytics Property ID (like UA-xxxxx)'
+    name: 'analyticsCodeBlock',
+    label: 'Custom code',
+    textarea: true,
   },
+
+  {
+    type: 'string',
+    name: 'analyticsIdentifier',
+    label: 'Google Analytics Property ID (like UA-xxxxx or G-xxxxx)'
+  },
+
   {
     type: 'string',
     name: 'tagmanager',